src/create_producer/requirements.txt
confluent-kafka==1.4.0 

#kafka-python